Part 3 - ReduxSaga
In this module we will see how to intercept requests with ReduxSaga, in this case we will be using to check the product quantity in stock. So our frontend can be more dynamic and intuitive, allowing the user to add products to their carts only if its available in stock.